requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.0
anyio==3.7.1
motor==3.3.2
pymongo==4.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.32.4
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
slowapi==0.1.9
typing-extensions==4.12.2